backend/requirements.txt
# Core LlamaIndex
llama-index-core
# LlamaIndex integrations
llama-index-llms-groq
llama-index-embeddings-huggingface
llama-index-vector-stores-postgres
llama-index-retrievers-bm25
# Embeddings runtime
sentence-transformers
# PostgreSQL / pgvector
psycopg2-binary
pgvector
sqlalchemy
# ChromaDB fallback
chromadb
llama-index-vector-stores-chroma
# API
fastapi
uvicorn[standard]
# Utilities
python-dotenv
pypdf
# BM25 dependency
rank-bm25